requirements.txt
alembic==1.6.2
asgiref==3.3.4
astroid==2.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.3
django-filter==2.4.0
django-restframework==0.0.1
djangorestframework==3.12.4
dodgy==0.2.1
flake8==3.8.4
flake8-polyfill==1.0.2
greenlet==1.1.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
pep8-naming==0.10.0
Pillow==8.1.2
prospector==1.3.1
protobuf==3.12.4
psutil==5.8.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pydocstyle==6.1.1
pyflakes==2.2.0
pylint==2.5.3
pylint-celery==0.3
pylint-django==2.1.0
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requirements-detector==0.7
setoptconf==0.2.0
six==1.16.0
snowballstemmer==2.1.0
SQLAlchemy==1.4.18
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
wrapt==1.12.1
